requirements.txt
Django==4.0.5
psycopg2-binary==2.9.3
djangorestframework==3.13.1
pytz==2022.1
pytest==7.1.2
pytest-django==4.5.2